+++ src/api/endpoint/chat.py
@@ -18,7 +18,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 
 from src.api.utils import error_response
 from src.core.database import get_db
-from src.core.models import Country
+from src.models import Country
 from src.schemas.chat import ChatRequest, ChatResult
 from src.schemas.common import CommonResponse
 from src.services.chat_service import generate_answer
