+++ src/api/endpoint/compare.py
@@ -7,7 +7,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 
 from src.api.utils import error_response
 from src.core.database import get_db
-from src.core.models import Country
+from src.models import Country
 from src.schemas.common import CommonResponse
 from src.schemas.compare import CompareRequest, CompareResult
 from src.services.compare_service import compare_laws
