+++ src/models/__init__.py
@@ -1,0 +1,5 @@
+from src.models.country import Country
+from src.models.law import Law
+from src.models.risk import Risk, RiskList
+
+__all__ = ["Country", "Law", "Risk", "RiskList"]
